requirements.txt
beautifulsoup4
discord.py
psycopg2-binary
requests